refactor(scripting): extract Lua worldgen bridge logic to worldgen_bridge.rs
